fix: attitude protection in steep turns
The attitude protection now also works in steep turns. Up to now it
was possible to override the -15°/+30° limitation in those situations.
The reason is the limitation of the curve compensation (> 33°), which
needs to be taken into account for the attitude limitation. The minimum
needs to be not the limited curve compensation but the non-limited
compensation.